Module stark_evm_adapter::annotation_parser

Structs

AnnotatedProof

AnnotatedProof maps annotated proof json file which contains the original proof and the annotations generated by verifier of stone-prover

SplitProofs

SplitProofs maps the split proof json file which contains the main proof and the merkle statements

Functions

split_fri_merkle_statements

This is the main function to use to split an AnnotatedProof file into a SplitProofs file. This processes the annotations of the original proof, and extra annotations prepared by a verifier, and returns a shortened proof (with merkle decommitments removed) in binary form, and a dictionary merkles_statements from the names of the merkle commitments used in the proof to dictionaries containing all data necessary for registering them in the Merkle Fact Registry.