Crate stark_evm_adapter

STARK-EVM adapter

stark-evm-adapter is a library that provides a set of utilities to parse and manipulate the output of the STARK stone proof. Specifically, the library can be used to generate a “split proof”, which is necessary for proofs to be verified on Ethereum.

[dependencies]
stark-evm-adapter = "0.1.3"

Example

use stark_evm_adapter::annotated_proof::AnnotatedProof;
use stark_evm_adapter::annotation_parser::split_fri_merkle_statements;

// read an annotated proof
let file = std::fs::File::open("tests/fixtures/annotated_proof.json").unwrap();
let reader = std::io::BufReader::new(file);
let annotated_proof: AnnotatedProof = serde_json::from_reader(reader).unwrap();

// split the proof
let split_proofs = split_fri_merkle_statements(annotated_proof).unwrap();
println!("{}", serde_json::to_string_pretty(&split_proofs).unwrap());

// For how to submit the split proofs to the L1 EVM verifier, please refer to the demo: https://github.com/zksecurity/stark-evm-adapter/blob/8af44a0aa61c89e36a08261320f234709e99ed71/examples/verify_stone_proof.rs#L18

Note that the annotated proof file, annotated_proof.json, can be generated using this CLI tool.

CLI

Installation

cargo install stark_evm_adapter

Usage

stark_evm_adapter --help

To generate an annotated proof based on the outputs of the stone-prover:

stark_evm_adapter gen-annotated-proof \
    --stone-proof-file tests/fixtures/stone_proof.json \
    --stone-annotation-file tests/fixtures/stone_proof_annotation.txt \
    --stone-extra-annotation-file tests/fixtures/stone_proof_annotation_extra.txt \
    --output annotated_proof.json

The annotated proof originates from 3 file outputs of the stone-prover.

Once you have this annotated proof, you can use it to generate the split proofs and submit them to the L1 EVM verifier. Please refer to the example demo

stone_proof.json comes from the cpu_air_prover command, while the annotation files come from the cpu_air_verifier command with arguments annotation_file and extra_output_file.

DEMO

You can use docker to run the demo. The demo will generate split proofs from an annotated proof and submit them to the L1 EVM verifier.

First, build the docker image:

docker build -t stark-evm-adapter .

Then, run the demo script:

docker run -it -e "MAINNET_RPC=******" stark-evm-adapter

Note that you will need to set the environment variable MAINNET_RPC to the RPC endpoint of the Ethereum mainnet. For example, you can use Infura to get the RPC endpoint. This demo code will automatically make a fork of the mainnet and submit the split proofs to the L1 EVM verifier on the forked chain.

Modules

annotated_proof

annotation_parser

errors

fri_merkle_statement

merkle_statement

oods_statement

Functions

default_prime

Default prime field for cairo. This prime will be used when modular operations are needed.

Type Aliases

ContractFunctionCall

A type alias for ethers contract function call.